Fix build warnings: data-infrastructure (5->0), testing-validation (2->0), data-interfaces-adapters improvements
- Fixed all warnings in data-infrastructure (unused client variables)
- Fixed all warnings in testing-validation (unreachable code)
- Reduced warnings in data-interfaces-adapters (lib 20->6, bin 14->11)
- Fixed duplicate key error in agent-constitutional-council Cargo.toml
- Added #[allow(dead_code)] to reserved fields and methods

+++ iterations/v3/testing-validation/src/scenarios/claim_verification.rs
@@ -31,7 +31,7 @@ pub async fn run_claim_verification_test(
     match test_claim_extraction(env, services).await {
         Ok(result) => {
             metrics.claims_extracted += result.claims_extracted;
-            metrics.disambiguations_resolved += result.disambiguations_resolved;
+            metrics.disambiguations_resolved += result.disambiguations_resolved as usize;
             if !result.passed {
                 passed = false;
                 errors.push(format!("Claim extraction failed: {}", result.error.unwrap_or_default()));
@@ -78,7 +78,7 @@ pub async fn run_claim_verification_test(
     // Test 4: Contextual disambiguation
     match test_contextual_disambiguation(env, services).await {
         Ok(result) => {
-            metrics.disambiguations_resolved += result.disambiguations_resolved;
+            metrics.disambiguations_resolved += result.disambiguations_resolved as usize;
             if !result.passed {
                 passed = false;
                 errors.push(format!("Contextual disambiguation failed: {}", result.error.unwrap_or_default()));
